+++ samples/SherlogSamples/GlobalUsings.cs
@@ -1,3 +1,4 @@
+global using System.Globalization;
 global using System.Net;
 global using Sherlog;
 global using Sherlog.Appenders;